requirements/unsloth_requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.1.31
charset-normalizer==3.4.1
cut-cross-entropy==25.1.1
datasets==3.5.0
diffusers==0.32.2
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
filelock==3.18.0
flash_attn==2.7.4.post1
frozenlist==1.5.0
fsspec==2024.12.0
hf_transfer==0.1.9
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==11.1.0
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==78.1.0
shtab==1.7.1
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
triton==3.2.0
trl==0.15.2
typeguard==4.4.2
typing_extensions==4.13.1
tyro==0.9.18
tzdata==2025.2
unsloth @ git+https://github.com/unslothai/unsloth.git@c9b9a366e7a6110f9d58d5ed8db6bd27bc97fb71
unsloth_zoo==2025.3.17
urllib3==2.3.0
wheel==0.45.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
